layout/inspector/tests/test_getComputationStepsSupportedCSSFunctions.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test InspectorUtils.getComputationStepsSupportedCSSFunctions</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<code>InspectorUtils.getComputationStepsSupportedCSSFunctions</code>

<script>
add_task(async function() {
  const { InspectorUtils } = SpecialPowers;

  const supportedFunctions = InspectorUtils.getComputationStepsSupportedCSSFunctions();
  const test = (functionName, included) => {
    is(
      supportedFunctions.includes(functionName),
      included,
      `"${functionName}" is${included ? "": " not"} returned by getComputationStepsSupportedCSSFunctions`
    );
  }

  test("calc", true);
  test("var", true);
  test("attr", true);
  test("env", true);
  test("sin", true);
  test("sibling-count", true);
  test("sibling-index", true);
  test("linear-gradient", false);
});
</script>
</body>
</html>
